requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.11
alembic==1.13.1
pydantic>=2.6,<3.0
pydantic-settings==2.2.1
jinja2==3.1.3
python-multipart==0.0.9
reportlab==4.0.8
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
cryptography==46.0.4
